Commit 07/01/2025
feat: 사용자 등록 기능 구현
 - 사용자 정보(UserDto)를 받아서 User 객체로 변환 후 등록
 - 중복된 사용자명과 이메일 체크 기능 추가

+++ ReadNSpeak/src/main/java/com/readnspeak/entity/User.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int user_id;
+    private long user_id;
 
     @Column(nullable = false, unique = true, length = 50)
     private String username;
@@ -34,11 +34,11 @@ public class User {
         USER, ADMIN
     }
 
-	public int getUser_id() {
+	public long getUser_id() {
 		return user_id;
 	}
 
-	public void setUser_id(int user_id) {
+	public void setUser_id(long user_id) {
 		this.user_id = user_id;
 	}
